Web App Beginning Front-end Files
Web app form interface to begin allowing the user to specify information that will ultimately help the application form and send over the user's intended query to be executed in the database. 

Right now, the table chosen is being stored in a particular variable, depending on whether the query is simple or complex.

Logic needs to be implemented that takes the table selected and based on the table, provides a listing of all the table's attributes for the user's use depending on the type of desired query (e.g. one involving a select clause). 

These two aspects (i.e. the table and the attribute(s)) will allow for the formation of a query where the select (i.e. attribute(s) are known), the from (i.e. table) is known, and further specifications (e.g. the attribute to join the tables on for a where for instance) will all be known. 

The current index.php has little space between the user's table selection and the submit button. 

A lot more content (e.g. the other are