Transactions Worked on, sample files to test
Transactions have been included in the project to account for cases where a CSV file has multiple statements and not all of them are valid. The goal here is to have it such that whenever a user uploads a CSV file to the program, either all of the file's statements get executed or none of them do. 

It can likely be further tested, but it appears functional at this time. 

The two CSV files, sample2.csv and sample3.csv were used to test that the transactional feature is at least functional at this time. 

1. To test, go through the normal procedures to access the Java program. 

2. Then, when prompted for to type in a CSV file name, first enter the sample2.csv file. This file has an insertion statement that is 
    invalid. 

3. When this file name is entered, the program should recognize that it contains an error and display a message indicating that the 
    transaction could not actually be committed, so none of the other insertion statements, regard