File to Connect Java Code to SQLite3 DB
Hey everyone, here's a file that contains code that seems like it could helpful to us in establishing a connection to our SQLite3 DB once we have the DB created in SQLIte3. 

Here is the source for the code: http://www.sqlitetutorial.net/sqlite-java/sqlite-jdbc-driver/

The webpage references where to download the SQLite JDBC Driver depending on the user's platform (e.g. Windows, Mac OS X, Linux, 32-bit or 64-bit).